+++ core/src/chr/gl/draw/Sprite.h
@@ -24,19 +24,19 @@ namespace chr
         Sprite& setAnchor(const glm::vec2 &anchor);
         Sprite& setAnchor(float x, float y);
 
-        template<int V = XYZ, typename I = GLushort>
+        template<int V = XYZ, typename I = GLuint>
         void append(IndexedVertexBatch<V,I> &batch, const Matrix &matrix, float x = 0, float y = 0) const;
 
-        template<int V = XYZ, typename I = GLushort>
+        template<int V = XYZ, typename I = GLuint>
         void append(IndexedVertexBatch<V,I> &batch, float x = 0, float y = 0) const;
 
-        template<int V = XYZ, typename I = GLushort>
+        template<int V = XYZ, typename I = GLuint>
         inline void append(IndexedVertexBatch<V,I> &batch, const Matrix &matrix, const glm::vec2 &position) const
         {
           append(batch, matrix, position.x, position.y);
         };
 
-        template<int V = XYZ, typename I = GLushort>
+        template<int V = XYZ, typename I = GLuint>
         inline void append(IndexedVertexBatch<V,I> &batch, const glm::vec2 &position) const
         {
           append(batch, position.x, position.y);
